svr: route: move all blockchain logic out of route
