+++ server/routes/blocks.test.js
@@ -6,6 +6,6 @@ suite('route: blocks', () => {
     const res = await testData.users.u1Request.get('/blocks/mine')
     assert(res.status === 200)
     assert(res.data.difficulty >= 1)
-    assert(res.data.str.length > 50)
+    assert(res.data.headerAsString.length > 50)
   })
 })
